# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart Cafe Manager API",
    "version": "1.0.0",
    "description": "REST API for the Smart Cafe Manager multi-tenant platform — Next.js App Router Route Handlers backed by Supabase. This spec is generated from `@swagger` JSDoc blocks on route handlers by scripts/generateSwagger.js; view it rendered at /api-docs. Only a representative subset of routes carries `@swagger` annotations so far (see the plan doc, Phase 11) — extend the pattern to the rest incrementally by adding the same style of block above any other exported GET/POST/PATCH/DELETE handler."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as this app"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT issued by /api/auth/staff-login, /api/auth/customer-login, or /api/auth/register. Send as `Authorization: Bearer <token>`."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "hiveId": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "master_admin",
              "owner",
              "manager",
              "cook",
              "waiter",
              "customer"
            ]
          },
          "authProvider": {
            "type": "string"
          },
          "spaceId": {
            "type": "string",
            "nullable": true
          },
          "preferences": {
            "type": "object",
            "properties": {
              "orderCount": {
                "type": "integer"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MenuItemOptionChoice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "priceDelta": {
            "type": "number"
          },
          "isDefault": {
            "type": "boolean"
          }
        }
      },
      "MenuItemOptionGroup": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "selectionType": {
            "type": "string"
          },
          "isRequired": {
            "type": "boolean"
          },
          "choices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MenuItemOptionChoice"
            }
          }
        }
      },
      "MenuItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "isAvailable": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "optionGroups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MenuItemOptionGroup"
            }
          }
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "menuItem": {
            "$ref": "#/components/schemas/MenuItem"
          },
          "quantity": {
            "type": "integer"
          },
          "priceAtPurchase": {
            "type": "number"
          },
          "selectedOptions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "choiceId": {
                  "type": "string"
                },
                "groupLabel": {
                  "type": "string"
                },
                "choiceLabel": {
                  "type": "string"
                },
                "priceDelta": {
                  "type": "number"
                }
              }
            }
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "totalAmount": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "preparing",
              "ready",
              "completed",
              "cancelled"
            ]
          },
          "orderTime": {
            "type": "string",
            "format": "date-time"
          },
          "mealType": {
            "type": "string",
            "enum": [
              "breakfast",
              "lunch",
              "dinner",
              "snack"
            ]
          },
          "orderType": {
            "type": "string",
            "enum": [
              "pickup",
              "dine_in",
              "delivery"
            ]
          },
          "tableNumber": {
            "type": "string",
            "nullable": true
          },
          "deliveryAddress": {
            "type": "string",
            "nullable": true
          },
          "deliveryLat": {
            "type": "number",
            "nullable": true
          },
          "deliveryLng": {
            "type": "number",
            "nullable": true
          },
          "deliveryPincode": {
            "type": "string",
            "nullable": true
          },
          "assignedCookId": {
            "type": "string",
            "nullable": true
          },
          "assignedRiderId": {
            "type": "string",
            "nullable": true
          },
          "deliveryStatus": {
            "type": "string",
            "nullable": true
          },
          "spaceId": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/orders/history": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "List orders, scoped by the caller role",
        "description": "Customer sees only their own orders; cook sees the unclaimed queue by default or their own claims with scope=mine; waiter sees the ready-for-delivery queue by default; manager/owner see everything. Requires a bearer token.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "fromDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "toDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "mealType",
            "schema": {
              "type": "string",
              "enum": [
                "breakfast",
                "lunch",
                "dinner",
                "snack"
              ]
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Overrides the role default queue filter"
          },
          {
            "in": "query",
            "name": "scope",
            "schema": {
              "type": "string",
              "enum": [
                "mine"
              ]
            },
            "description": "Cook only: restrict to orders assigned to the caller"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated order history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orders": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Order"
                      }
                    },
                    "totalPages": {
                      "type": "integer"
                    },
                    "currentPage": {
                      "type": "integer"
                    },
                    "totalOrders": {
                      "type": "integer"
                    },
                    "hasNextPage": {
                      "type": "boolean"
                    },
                    "hasPrevPage": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/menu": {
      "get": {
        "tags": [
          "Menu"
        ],
        "summary": "List menu items",
        "description": "Scoped to the caller organization when a valid bearer token is present. Public (no 401) when no token is sent - there is no per-organization anonymous menu page yet.",
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Menu items, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MenuItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Menu"
        ],
        "summary": "Create a menu item",
        "description": "Manager only. Requires a bearer token.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "category",
                  "price"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "description": {
                    "type": "string"
                  },
                  "imageUrl": {
                    "type": "string"
                  },
                  "isAvailable": {
                    "type": "boolean",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Menu item created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItem"
                }
              }
            }
          },
          "400": {
            "description": "Missing name, category, or price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not a Manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}